requirements.txt
opencv-python~=4.5.5.64
ffmpeg-python
PyYAML~=6.0
pygame~=2.1.2
tqdm~=4.63.1
py2exe~=0.11.1.0